Migrate the primary layout to the Hugo world
We're migrating https://git-scm.com/ to a static site generated using
Hugo, so let's move the layout to the appropriate place.

The layout will be adjusted in the following commit, to allow help Git's
rename detection (which would otherwise be confused into believing that
we deleted a file and added a new file elsewhere, which is not at all
what we did here).

Signed-off-by: Matt Burke <spraints@gmail.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>