Move the "references" landing page to `layouts/partials/`, in preparation for using it as a Hugo partial template.
This partial template will be subsequently modified so as to render
correctly via Hugo.

Doing this in two separate commits helps overcome the shortcomings of
Git's rename detection, which would be foiled otherwise into believing
that the original file was deleted and the partial template was conjured
out of thin air.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>